requirements.txt
pandas
numpy
streamlit
plotly
scikit-learn
fpdf
requests
paho-mqtt
